requirements.txt
# requirements_streamlit.txt
# Dépendances supplémentaires pour l'interface Streamlit

# Interface web
streamlit>=1.28.0

# Visualisations
plotly>=5.17.0
pandas>=2.0.0

# Utilitaires pour l'interface
pillow>=10.0.0

# Dépendances manquantes identifiées
PyYAML>=6.0.0
requests>=2.28.0
selenium>=4.0.0
beautifulsoup4>=4.11.0
lxml>=4.9.0